+++ tests/testthat/test-rd-params.R
@@ -1,20 +1,43 @@
 test_that("@param documents arguments", {
-  out <- roc_proc_text(rd_roclet(), "
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
     #' A
     #' @param a first
     #' @param z last
     a <- function(a=1, z=2) {}
-  ")[[1]]
+  "
+  )[[1]]
 
   expect_equal(out$get_value("param"), c(a = "first", z = "last"))
 })
 
+test_that("backtick-quoted @param names are parsed correctly (#1696)", {
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
+    #' A
+    #' @param `arg 1` first
+    #' @param `arg 2` last
+    a <- function(`arg 1` = 1, `arg 2` = 2) {}
+  "
+  )[[1]]
+
+  expect_equal(
+    out$get_value("param"),
+    c("arg 1" = "first", "arg 2" = "last")
+  )
+})
+
 test_that("grouped args get spaces", {
-  out <- roc_proc_text(rd_roclet(), "
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
     #' A
     #' @param a,z Two arguments
     a <- function(a=1, z=2) {}
-  ")[[1]]
+  "
+  )[[1]]
   expect_match(out$get_rd("param"), "a, z")
 })
 
@@ -30,36 +53,45 @@ test_that("empty @param generates warning", {
 })
 
 test_that("data objects don't get params", {
-  out <- roc_proc_text(rd_roclet(), "
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
     #' x
     #' @rdname xy
     x <- 'x'
-  ")[[1]]
+  "
+  )[[1]]
   expect_equal(out$get_value("param"), NULL)
 })
 
 test_that("arguments ordered by usage", {
-  out <- roc_proc_text(rd_roclet(), "
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
     #' A
     #'
     #' @param y Y
     #' @param x X
     #' @rdname rd
     a <- function(x, y) {}
-  ")[[1]]
+  "
+  )[[1]]
 
   expect_named(out$get_value("param"), c("x", "y"))
 })
 
 test_that("multiple arguments ordered by first", {
-  out <- roc_proc_text(rd_roclet(), "
+  out <- roc_proc_text(
+    rd_roclet(),
+    "
     #' Title
     #'
     #' @param y Y
     #' @param x,z X,Z
     #' @param w W
     b <- function(x, y, z, w) {}
-  ")[[1]]
+  "
+  )[[1]]
 
   expect_named(out$get_value("param"), c("x,z", "y", "w"))
 })